# 03_openapi_original.yaml
openapi: 3.0.3
info:
  title: Kaveri Stays API
  version: 1.0.0
  description: API for Kaveri Stays
servers:
  - url: http://127.0.0.1:8000
...
paths:
  /auth/register:
    post:
      tags: [Authentication]
      summary: Register a guest
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201': {description: Guest account created}
        '400': {description: Invalid request}
        '409': {description: Account already exists}
  /auth/login:
    post:
      tags: [Authentication]
      summary: Login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401': {description: Invalid credentials or disabled account}
  /auth/refresh:
    post:
      tags: [Authentication]
      summary: Refresh access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200': {description: Tokens refreshed}
        '401': {description: Invalid or revoked refresh token}
  /auth/logout:
    post:
      tags: [Authentication]
      summary: Logout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '204': {description: Logged out}
        '401': {description: Missing or invalid access token}
...
  /me:
    get:
      tags: [Me]
      summary: Get current account
      responses:
        '200': {description: Current account}
        '401': {description: Missing or invalid token}
components:
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    RegisterRequest:
      type: object
      required: [full_name, email, password]
      properties:
        full_name: {type: string}
        email: {type: string, format: email}
        password: {type: string, minLength: 8}
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: {type: string, format: email}
        password: {type: string}
    LoginResponse:
      type: object
      properties:
        message: {type: string}
        access_token: {type: string}
        token_type: {type: string, example: bearer}
        account_id: {type: integer}
        guest_id: {type: integer, nullable: true}
        role: {type: string}
        refresh_token: {type: string}
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token: {type: string}
...